Split hashing tests
